+++ docs/examples/layout/slider.py
@@ -12,8 +12,8 @@ curve = Bezier([[0, 0, 0], [3, 6, 0], [5, -3, 0], [10, 0, 0]])
 viewer = Viewer(rendermode="shaded", width=1600, height=900)
 
 
-pointobj: PointObject = viewer.add(Point(*curve.point_at(0)), pointssize=20, pointscolor=Color.red(), show_points=True)  # type: ignore
-curveobj = viewer.add(Polyline(curve.to_polyline()), lineswidth=2, linescolor=Color.blue())
+pointobj: PointObject = viewer.scene.add(Point(*curve.point_at(0)), pointssize=20, pointscolor=Color.red(), show_points=True)  # type: ignore
+curveobj = viewer.scene.add(Polyline(curve.to_polyline()), lineswidth=2, linescolor=Color.blue())
 
 
 def slide(value):